Revert "Moved social preview forms to components"
This reverts commit 1aac196cee01fbccca06647671ee4629f6bdb99f.
