+++ packages/components/src/index.js
@@ -42,7 +42,6 @@ export { default as Icon } from "./Icon";
 export { default as IconButtonToggle } from "./IconButtonToggle.js";
 export { default as IFrame } from "./IFrame";
 export { default as Input } from "./input/Input";
-export { default as ImageSelect } from "./ImageSelect.js";
 export { default as KeywordSuggestions } from "./KeywordSuggestions";
 export { default as Label } from "./Label";
 export { default as LanguageNotice, languageNoticePropType } from "./LanguageNotice";
@@ -53,7 +52,6 @@ export { default as Paper } from "./Paper";
 export { default as ProgressBar } from "./ProgressBar";
 export { default as Section } from "./Section";
 export { SectionTitle } from "./SectionTitle";
-export { default as SocialMetadataPreviewForm } from "./SocialMetadataPreviewForm.js";
 export { default as ScoreAssessments } from "./ScoreAssessments";
 export { default as StackedProgressBar } from "./StackedProgressBar";
 export { default as SvgIcon, icons } from "./SvgIcon";
