+++ packages/social-metadata-previews/tests/shared/ImageUploadTest.js
@@ -2,7 +2,7 @@
 import React from "react";
 import renderer from "react-test-renderer";
 /* Internal dependencies */
-import ImageUpload from "@yoast/components/src/ImageSelect";
+import ImageUpload from "../../src/shared/ImageUpload";
 
 describe( "<ImageUpload />", () => {
 	it( "renders", () => {
